hood: another one

+++ toontown/hood/ToonHood.py
@@ -231,6 +231,8 @@ class ToonHood(Hood):
                                         'state': currentButterfly.getState()})
 
     def unloadButterflies(self):
+        if not hasattr(self, "butterflies"):
+            return
         for butterfly in self.butterflies:
             butterfly.disable()
             butterfly.delete()
